pyproject.toml
[tool.poetry]
name = "apirestfull_curso"
version = "0.1.0"
description = ""
authors = ["MarcoAntonioGritti <m9946658@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4"
flask-sqlalchemy = "3.1.1"
flask-migrate = "4.1.0"
flask-jwt-extended = "4.7.1"
flask = "3.1.0"
jinja2 = ">=3.1.5,<4.0.0"
itsdangerous = ">=2.2.0,<3.0.0"
gunicorn = "^23.0.0"
psycopg = "^3.2.4"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
flask-bcrypt = "^1.0.1"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "^1.4.1"

[tool.poetry.group.dev.dependencies]
poetry = ">=2.0.1,<3.0.0"
pytest = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
